+++ src/test_files/iso.input
@@ -1,1 +1,1 @@
-זרו
+זרו45eeaoy
