+++ src/test_files/random_bytes.input
