--- src2/test.sh
@@ -1,47 +1,0 @@
-#!/usr/bin/env bash
-
-# Exit immediately if any command below fails.
-set -e
-
-make
-
-
-echo "Generating a test_files directory.."
-mkdir -p test_files
-rm -f test_files/*
-
-
-echo "Generating test files.."
-printf "Hello, World!\n" > test_files/ascii.input
-printf "Hello, World!" > test_files/ascii2.input
-printf "Hello,\x00\x00\x00\x00 World!\n" > test_files/data.input
-printf "" > test_files/empty.input
-printf "Hej med æøå og emojis" > test_files/utf8.input
-printf "æøå45eeaoy" | iconv -f UTF-8 -t ISO-8859-1 > test_files/iso.input
-head -c 100 /dev/urandom > test_files/random_bytes.input
-
-
-### TODO: Generate more test files ###
-
-
-echo "Running the tests.."
-exitcode=0
-for f in test_files/*.input
-do
-  echo ">>> Testing ${f}.."
-  file    ${f} | sed -e 's/ASCII text.*/ASCII text/' \
-                         -e 's/UTF-8 Unicode text.*/UTF-8 Unicode text/' \
-                         -e 's/ISO-8859 text.*/ISO-8859 text/' \
-                         -e 's/writable, regular file, no read permission/cannot determine (Permission denied)/' \
-                         > "${f}.expected"
-  ./file  "${f}" > "${f}.actual"
-
-  if ! diff -u "${f}.expected" "${f}.actual"
-  then
-    echo ">>> Failed :-("
-    exitcode=1
-  else
-    echo ">>> Success :-)"
-  fi
-done
-exit $exitcode
