--- src2/test_files/ascii.input
@@ -1,1 +1,0 @@
-Hello, World!
