--- src2/test_files/ascii.input.actual
@@ -1,1 +1,0 @@
-test_files/ascii.input: ASCII text
