--- src2/test_files/ascii.input.expected
@@ -1,1 +1,0 @@
-test_files/ascii.input: ASCII text
