--- src2/test_files/ascii2.input
@@ -1,1 +1,0 @@
-Hello, World!
