--- src2/test_files/ascii2.input.expected
@@ -1,1 +1,0 @@
-test_files/ascii2.input: ASCII text
