--- src2/test_files/data.input.actual
@@ -1,1 +1,0 @@
-test_files/data.input: data
