--- src2/test_files/data.input.expected
@@ -1,1 +1,0 @@
-test_files/data.input: data
