--- src2/test_files/empty.input.actual
@@ -1,1 +1,0 @@
-test_files/empty.input: empty
