--- src2/test_files/iso.input
@@ -1,1 +1,0 @@
-זרו45eeaoy
