--- src2/test_files/iso.input.actual
@@ -1,1 +1,0 @@
-test_files/iso.input: ISO-8859 text
