--- src2/test_files/utf8.input
@@ -1,1 +1,0 @@
-Hej med æøå og emojis
