--- src2/test_files/utf8.input.expected
@@ -1,1 +1,0 @@
-test_files/utf8.input: Unicode text, UTF-8 text, with no line terminators
