--- src2/utf8.input
@@ -1,1 +1,0 @@
-Hej med æøå og emojis
